feat(michelson-encoder): Support pair as key in map and big map
Create a new MichelsonMap class that allow initialization of storage which have map that expect pair
as key. MichelsonMap is also used to represent map when reading from storage. This feature also make
the pair token comparable in Taquito as well as the boolean token.

BREAKING CHANGE: Using object literal to initialize map is not working anymore and map are
represented using MichelsonMap when reading storage

closes #269