fix(estimator) Fix a bug where explicit limits were ignored by the estimator
Description of the issue:

Setting the storageLimit to 0 on an origination did not threw an error and did a dry run with the maximum storageLimit (60000). Also the actual operation sent by taquito
contained the explicit storageLimit instead of the that the user specified which lead to a failure on chain. So in this example the origination would have been reported as valid to the user but rejected by the chain.

- Ensure that we respect the explicit limit set by the user while estimating
- Add unit test coverage and integration test coverage for this issue