fix(taquito-signer): ensure correct byte is used to determine parity
We were looking at the byte at position 31 of the "y" coordinate to determine which prefix to use (3
if odd or 2 if even) when deriving the public key. The bug happened in cases where the "y"
coordinate was shorter than 32 bytes. The value at position 31 was undefined, so the prefix "2" was
always used enven if the "y" coordinate was odd.

fix #848