feat(beacon-wallet): update beacon-sdk to version 2.3.0-beta.1
A call is made to the showPrepare method of the dAppClient while Taquito prepares the operation.
This will show a toast that says "Preparing request..."

BREAKING CHANGE: The wallet interface has been modified. The mapTransferParamsToWalletParams method
accepts a function that returns the WalletTransferParams as a parameter instead of directly taking
the WalletTransferParams as a parameter. Same change for mapOriginateParamsToWalletParams and
mapDelegateParamsToWalletParams.