fix(contractabstraction): getting ContractAbstraction for large contracts takes too much resources
ParameterSchema.ExtractSignature was used during ContractAbstraction instanciation to determine
which entry point are lambda views. For particular contracts having nested pair, or and option
types, ExtractSignature takes a lot of ressources (memory) when generating all possible signatures.
The usage of ExtractSignature was unnecessary as the entry point type can be directly parsed to
determine if it is a view.