fix(ithacanet): block needs to be HEAD~2
When preparing an operation the block was set to head. In the Ithaca protocol, the block needs to be
HEAD~2. Setting the branch field to HEAD or HEAD~1 may result in the operation not being included
because it will not be anchored on a block belonging to the chain. (The blocks at the current and
previous levels are not necessary final.)

fix #1230