refactor(operation confirmation): use the SubscribeProvider set on the TezosToolkit for confirmation
Use the SubscribeProvider set on the TezosToolkit in the operation confirmation step. It is set by
default to an instance of PollingSubscribeProvider which polls on the RPC for head block.

BREAKING CHANGE: The confirmationPollingIntervalSecond and the ConfigStreamer are removed from the
TezosToolkit. Configuration for the PollingSubscribeProvider need to be pass in its constructor.

re #1389