Read Interface (#1397)
* Interfaces read provider initial commit

* RpcReadAdapter

* refactor: Make the read calls using the configured TzReadProvider instead of the RPC (#1400)

* refactor: replace rpc by readProvider for the read calls

Changes are required in the codebase to read data from the configured ReadProvider.

#1389

* refactor(operation confirmation): use the SubscribeProvider set on the TezosToolkit for confirmation

Use the SubscribeProvider set on the TezosToolkit in the operation confirmation step. It is set by
default to an instance of PollingSubscribeProvider which polls on the RPC for the head block.

BREAKING CHANGE: The confirmationPollingIntervalSecond and the ConfigStreamer are removed from the
TezosToolkit. Configuration for the PollingSubscribeProvider needs to be passed in its constructor.

re #1389

* Fix config streamer in examples from the example folder

* feat(read interface): added missing methods on the read interface