feat(estimate): fix estimate to work with wallet (#1438)
* feat(estimate): fix estimate to work with wallet

Estimation is done with the configured signer. If there is no configured sihner, estimation is done
using the public key hash of the configured wallet. Taquito does not have access to the public key
of the wallet (only the pkh), so wont estimate reveal operation. If the account is unrevealed, an
exception is throwned.

re #1387

* Renamed function as suggested in review