fix(contract api): fetch missed blocks during operation confirmation
During the confirmation of an operation (step where we poll on the head block until the injected
operation is found), there was no mecanism to detect missed blocks. If the operation was included in
a skipped block, the operation was never found. Made a change to fetch the skipped blocks and verify
if the operation is inside those.

fix #1783