fix(wallet and contract APIs): fetch missed blocks during operation confirmation (#2003)
* fix(walletapi): fetch missed blocks during operation confirmation

During the confirmation of an operation (step where we poll on the head block until the injected
operation is found) Taquito was throwing an error MissedBlockDuringConfirmationError when blocks
were skipped. Made a change to fetch the skipped blocks and verify if the operation is inside
instead of throwing.

re #1783

* fix(contract api): fetch missed blocks during operation confirmation

During the confirmation of an operation (step where we poll on the head block until the injected
operation is found), there was no mecanism to detect missed blocks. If the operation was included in
a skipped block, the operation was never found. Made a change to fetch the skipped blocks and verify
if the operation is inside those.

fix #1783

* test: add unit tests

Co-authored-by: Michael Kernaghan <michaelkernaghan@ecadlabs.com>