test(michel-codec): added lima contracts for unit tests
There is no code change in the branch. I removed the contracts related to protocols 10 to 13 and
added contracts for the protocol 15. Some contracts are commented out in the contract_015.spec.ts
file as they are currently failing. Those contracts should be uncommented and fixed when working on
the issue #2075.

re $2075