fix(michelson-encoder): fix for MapTypecheckError
Quick fix for the method Schema.Typecheck that fails to parse map when its value is of type big_map.
Note that it is not the first bug happening with the Typecheck method: using "EncodeObject" to do
the validation is not always appropirate. I thik this part of the code should be rework in the
futur.

re #1762